Test code documentation improvements (#53)
* Added docstrings to test files where it was orginally missing
* Updated existing test case docstrings with minor changes
* Moved the different test classes in test_utils.py into separate files in the tests/utils/ module
* Moved the different test classes in test_types.py into separate files in the tests/types/ module
