+++ tests/cmd/outputs/test_json_output.py
@@ -1,9 +1,9 @@
-"""Test cases for the `TableOutput` static class
+"""Test cases for the JSONOutput static class
 
 Test Cases:
-    * `test_displayed_violation_output` tests that the expected status code
-       is returned when valid arguments are provided
-    * `test_json_output_invalid_args` tests that the relevant exception is
+    * `test_json_output` tests that the expected status code is returned
+       when valid arguments are provided
+    * `test_json_output_invalid_args` tests that the expected exception is
        raised when invalid arguments are provided
 """
 
@@ -16,15 +16,14 @@ from unittest.mock import patch
 from parameterized import parameterized
 
 from yamlator.cmd.outputs import SuccessCode
-from yamlator.cmd.outputs import TableOutput
+from yamlator.cmd.outputs import JSONOutput
 from yamlator.violations import RequiredViolation
-from yamlator.violations import Violation
 from yamlator.violations import TypeViolation
-from yamlator.violations import ViolationType
+from yamlator.violations import Violation
 
 
-class TestTableOutput(unittest.TestCase):
-    """Test the `TableOutput` display method"""
+class TestJSONOutput(unittest.TestCase):
+    """Test the JSONOutput display method"""
 
     @parameterized.expand([
         ('with_no_violations', [], SuccessCode.SUCCESS),
@@ -33,19 +32,19 @@ class TestTableOutput(unittest.TestCase):
             TypeViolation(key='number', parent='-', message='Invalid number')
         ], SuccessCode.ERR)
     ])
-    def test_displayed_violation_output(self, name: str,
-                                        violations: Iterator[ViolationType],
-                                        expected_status_code: int):
+    def test_json_output(self, name: str,
+                         violations: Iterator[Violation],
+                         expected_status_code: str):
         # Unused by test case, however is required by the parameterized library
         del name
 
         # Suppress the print statements
         with patch('sys.stdout', new=io.StringIO()):
-            status_code = TableOutput.display(violations)
+            status_code = JSONOutput.display(violations)
             self.assertEqual(expected_status_code, status_code)
 
     @parameterized.expand([
-        ('with_non_violations', None, ValueError)
+        ('with_none_violations', None, ValueError)
     ])
     def test_json_output_invalid_args(self, name: str,
                                       violations: Iterator[Violation],
@@ -54,7 +53,7 @@ class TestTableOutput(unittest.TestCase):
         del name
 
         with self.assertRaises(expected_exception):
-            TableOutput.display(violations)
+            JSONOutput.display(violations)
 
 
 if __name__ == '__main__':
