+++ tests/cmd/outputs/test_table_output.py
@@ -1,9 +1,9 @@
-"""Test cases for the `JSONOutput` static class
+"""Test cases for the TableOutput static class
 
 Test Cases:
-    * `test_json_output` tests that the expected status code is returned
-       when valid arguments are provided
-    * `test_json_output_invalid_args` tests that the relevant exception is
+    * `test_displayed_violation_output` tests that the expected status code
+       is returned when valid arguments are provided
+    * `test_json_output_invalid_args` tests that the expected exception is
        raised when invalid arguments are provided
 """
 
@@ -16,14 +16,15 @@ from unittest.mock import patch
 from parameterized import parameterized
 
 from yamlator.cmd.outputs import SuccessCode
-from yamlator.cmd.outputs import JSONOutput
+from yamlator.cmd.outputs import TableOutput
 from yamlator.violations import RequiredViolation
-from yamlator.violations import TypeViolation
 from yamlator.violations import Violation
+from yamlator.violations import TypeViolation
+from yamlator.violations import ViolationType
 
 
-class TestJSONOutput(unittest.TestCase):
-    """Test the JSONOutput display method"""
+class TestTableOutput(unittest.TestCase):
+    """Test the `TableOutput` display method"""
 
     @parameterized.expand([
         ('with_no_violations', [], SuccessCode.SUCCESS),
@@ -32,19 +33,19 @@ class TestJSONOutput(unittest.TestCase):
             TypeViolation(key='number', parent='-', message='Invalid number')
         ], SuccessCode.ERR)
     ])
-    def test_json_output(self, name: str,
-                         violations: Iterator[Violation],
-                         expected_status_code: str):
+    def test_displayed_violation_output(self, name: str,
+                                        violations: Iterator[ViolationType],
+                                        expected_status_code: int):
         # Unused by test case, however is required by the parameterized library
         del name
 
         # Suppress the print statements
         with patch('sys.stdout', new=io.StringIO()):
-            status_code = JSONOutput.display(violations)
+            status_code = TableOutput.display(violations)
             self.assertEqual(expected_status_code, status_code)
 
     @parameterized.expand([
-        ('with_none_violations', None, ValueError)
+        ('with_non_violations', None, ValueError)
     ])
     def test_json_output_invalid_args(self, name: str,
                                       violations: Iterator[Violation],
@@ -53,7 +54,7 @@ class TestJSONOutput(unittest.TestCase):
         del name
 
         with self.assertRaises(expected_exception):
-            JSONOutput.display(violations)
+            TableOutput.display(violations)
 
 
 if __name__ == '__main__':
