+++ tests/cmd/test_display_violations.py
@@ -1,10 +1,10 @@
-"""Test cases for the `display_violations` function
+"""Test cases for the display_violations function
 
 Test cases:
     * `test_display_violations` tests that the function will display the
        results when provided a valid display method
-    * `test_display_violations_invalid_params` tests that the expected
-       exception is raised when invalid parameters are provided.
+    * `test_display_violations_invalid_params` tests that a ValueError
+       exception is raised when invalid parameters are provided
 """
