+++ tests/parser/test_schema_transformer.py
@@ -1,58 +1,4 @@
-"""Test cases for the `SchemaTransformer`
-
-Test Cases:
-    * `test_required_rule` tests the transformation of a required rule token
-       into a rule object with required set to `true`
-    * `test_optional_rule` tests the transformation of a optional rule token
-       into a rule object with required set to `false`
-    * `test_ruleset` tests that given a set of rule tokens, a ruleset is
-       generated that contains all the specified rules
-    * `test_start` tests that all the rulesets and enums are sorted into
-       the correct container type, which can then be used by the Yamlator
-       validation chain
-    * `test_str_type` tests transforming a string token into a Yamlator
-       string rule type
-    * `test_int_type` tests transforming a int token into a Yamlator
-       int rule type
-    * `test_float_type` tests transforming a float token into a Yamlator
-       float rule type
-    * `test_list_type` tests transforming a list token into a Yamlator
-       list rule type
-    * `test_map_type` tests transforming a map token into a Yamlator
-       map rule type
-    * `test_any_type` tests transforming a any token into a Yamlator
-       any rule type
-    * `test_bool_type` tests transforming a bool token into a Yamlator
-       bool rule type
-    * `test_enum_item` tests transforming enum item tokens into a Yamlator
-       enum item
-    * `test_enum` tests that given a set of enum items a enum container
-       type is created
-    * `test_container_type` tests transforming a container Type
-       (ruleset / enum) type into the relevant rule type
-    * `test_container_type_construct_does_not_exist` tests transforming a
-       container type when the name of the given container is not found
-    * `test_regex_type` tests transforming regex tokens into a Yamlator
-       regex type
-    * `test_union_type` tests transforming a union token into a Union
-       rule type
-    * `test_union_type_with_nested_union` tests that a exception
-       is raised when a union contains a nested union
-    * `test_type` tests the type transformer returns the token it was
-       passed so the other transformer can successfully transform it
-    * `test_schema_entry` tests that given a set of rule tokens within
-       a schema block, a schema block is generated that contains all
-       the specified rules
-    * `test_strict_schema_entry` tests that given a set of rule tokens within
-       a schema block, a schema block is generated that contains all
-       the specified rules in strict mode
-    * `test_int_transform` tests transforming a int into an actual
-       integer value
-    * `test_float_transform` tests transforming a float into an actual
-       float value
-    * `test_escaped_string_transform` tests that a string is escaped
-       and any speech marks are removed
-"""
+"""Test cases for the SchemaTransformer"""
 
 
 import re
